requirements.txt
streamlit
deep-translator
python-docx
PyPDF2
pillow
pytesseract
